+++ packages/iles/tsup-cjs.config.ts
@@ -1,7 +1,7 @@
 import type { Options } from 'tsup'
 export const tsup: Options = {
   dts: true,
-  target: 'node14',
+  target: 'node18',
   splitting: false,
   sourcemap: false,
   format: ['cjs'],