+++ packages/iles/types/shared.d.ts
@@ -11,7 +11,7 @@ import type { Options as SvelteOptions } from '@sveltejs/vite-plugin-svelte'
 import type { PreactPluginOptions as PreactOptions } from '@preact/preset-vite'
 
 import type { Router, RouteRecordRaw, RouteMeta, RouterOptions as VueRouterOptions, RouteComponent, RouteRecordNormalized, RouteLocationNormalizedLoaded, RouteParams } from 'vue-router'
-import type { HeadClient, HeadObject } from '@vueuse/head'
+import type { HeadClient, HeadObject } from '@unhead/vue'
 import type { PagesApi, PagesOptions, PageFrontmatter, PageMeta } from '@islands/pages'
 export type { RawPageMatter, PageFrontmatter, PageMeta } from '@islands/pages'
 export type { OnLoadFn } from '@islands/hydration/dist/vanilla'