+++ packages/pwa/tsup.config.ts
@@ -2,7 +2,7 @@ import type { Options } from 'tsup'
 export const tsup: Options = {
   clean: true,
   dts: true,
-  target: 'node14',
+  target: 'node18',
   splitting: true,
   format: ['cjs'],
 }